Abstract the postgreslql setup in pyapps

+++ core/config/generator/pyapps_cfg.py
@@ -49,47 +49,56 @@ class PyAppsConfigGen:
             pyappsCfg.write()
             self._load_config()
 
-    def setup(self):
+    def setup_postgres(self):
         from pylabs.db import DBConnection
-        #create user with applicationname
-        if not q.system.unix.unixUserExists(self.appName):
-            q.system.unix.addSystemUser(self.appName)
-        if 'postgresql' in self.components:
-            postgres = q.manage.postgresql8
-            if self.appName not in postgres.cmdb.databases:
-                postgres.startChanges()
-                if not postgres.cmdb.initialized:
-                    postgres.cmdb.initialized = True
-                    postgres.cmdb.rootLogin = POSTGRESUSER
-                    postgres.cmdb.addLogin(POSTGRESUSER)
-                db = postgres.cmdb.addDatabase(self.appName, self.appName)
-                db.initDone = True
-                db.new = False
-                db.addACE(self.appName, '', q.enumerators.PostgresqlAccessRightType.WRITE)
-                postgres.cmdb.addLogin(self.appName,  type='host',
-                        cidr_address='127.0.0.1/32',database=self.appName)
-                postgres.save()
-                postgres.applyConfig()
-                q.manage.postgresql8.start()
-                conn = DBConnection.DBConnection('127.0.0.1', 'postgres', POSTGRESUSER)
-                createdb = True
-                if {'datname': self.appName} in conn.sqlexecute("select datname from pg_catalog.pg_database").dictresult():
-                    if q.console.askYesNo('Database already exists, overwrite existing database?'):
-                        q.cmdtools.postgresql8.dropdb(self.appName, POSTGRESUSER)
-                    else:
-                        createdb = False
-                if createdb:
-                    q.cmdtools.postgresql8.createdb(self.appName, POSTGRESUSER, self.appName)
-            dbconnections = q.config.getInifile('dbconnections')
-            section = "db_%s" % self.appName
-            if not dbconnections.checkSection(section):
-                dbconnections.addSection(section)
+        dbconnections = q.config.getInifile('dbconnections')
+        section = "db_%s" % self.appName
+        dbhost = '127.0.0.1'
+        if not dbconnections.checkSection(section):
+            dbconnections.addSection(section)
             dbconnections.addParam(section, 'dbtype', 'postgresql')
             dbconnections.addParam(section, 'dbserver', '127.0.0.1')
             dbconnections.addParam(section, 'dblogin', self.appName)
             dbconnections.addParam(section, 'dbpasswd', '')
             dbconnections.addParam(section, 'dbname', self.appName)
             dbconnections.write()
+        else:
+            dbhost = dbconnections.getValue(section, 'dbserver')
+
+        islocal = dbhost in ('localhost', '127.0.0.1')
+        postgres = q.manage.postgresql8
+        if islocal and self.appName not in postgres.cmdb.databases:
+            postgres.startChanges()
+            if not postgres.cmdb.initialized:
+                postgres.cmdb.initialized = True
+                postgres.cmdb.rootLogin = POSTGRESUSER
+                postgres.cmdb.addLogin(POSTGRESUSER)
+            db = postgres.cmdb.addDatabase(self.appName, self.appName)
+            db.initDone = True
+            db.new = False
+            db.addACE(self.appName, '', q.enumerators.PostgresqlAccessRightType.WRITE)
+            postgres.cmdb.addLogin(self.appName,  type='host',
+                    cidr_address='127.0.0.1/32',database=self.appName)
+            postgres.save()
+            postgres.applyConfig()
+            q.manage.postgresql8.start()
+            conn = DBConnection.DBConnection('127.0.0.1', 'postgres', POSTGRESUSER)
+            createdb = True
+            if {'datname': self.appName} in conn.sqlexecute("select datname from pg_catalog.pg_database").dictresult():
+                if q.console.askYesNo('Database already exists, overwrite existing database?'):
+                    q.cmdtools.postgresql8.dropdb(self.appName, POSTGRESUSER)
+                else:
+                    createdb = False
+            if createdb:
+                q.cmdtools.postgresql8.createdb(self.appName, POSTGRESUSER, self.appName)
+
+
+    def setup(self):
+        #create user with applicationname
+        if not q.system.unix.unixUserExists(self.appName):
+            q.system.unix.addSystemUser(self.appName)
+        if 'postgresql' in self.components:
+            self.setup_postgres()
         if 'wfe' in self.components:
             if self.appName not in q.manage.ejabberd.cmdb.hosts:
                 agent_cfg = AgentPyApps(self.appName)
